oracle-api/requirements.txt
flask>=3.0
web3>=6.15
python-dotenv>=1.0
